assets/requirements.txt
arrow
psutil
pywin32 | Windows
requests
winrt-Windows.UI.ViewManagement | Windows
winrt-Windows.UI | Windows
pygit2
cssutils
websockets
watchdog